+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased][]
+### Fixed
+- Item tooltip regression
+- Item embed being hidden until loaded
+- Embeds not showing api errors correctly
 
 ## [4.35.0][] - 2017-09-05
 ### Added