+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased][]
+
+## [4.35.1][] - 2017-09-05
 ### Fixed
 - Item tooltip regression
 - Item embed being hidden until loaded
@@ -991,7 +993,8 @@ At a later date a embed generator page will be made to help with the creation of
 ### Added
 - This CHANGELOG file to allow more insight to the changes made throughout the development of gw2armory.com
 
-[Unreleased]: https://github.com/madou/armory-react/compare/v4.35.0...HEAD
+[Unreleased]: https://github.com/madou/armory-react/compare/v4.35.1...HEAD
+[4.35.1]: https://github.com/madou/armory-react/compare/v4.35.0...v4.35.1
 [4.35.0]: https://github.com/madou/armory-react/compare/v4.34.0...v4.35.0
 [4.34.0]: https://github.com/madou/armory-react/compare/v4.33.13...v4.34.0
 [4.33.13]: https://github.com/madou/armory-react/compare/v4.33.12...v4.33.13